pyproject.toml
[project]
name = "novel-assistant-backend"
version = "0.1.0"
description = "Local SQLite API for the novel assistant."
license = "Apache-2.0"
requires-python = ">=3.9"
dependencies = [
    "llama-cpp-python>=0.3.20",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
]

[tool.ruff]
line-length = 100

[tool.pyright]
pythonVersion = "3.9"
typeCheckingMode = "basic"
